requirements.txt
kivy>=2.3.0
Pillow>=10.0.0
plyer>=2.1.0